app/bromonitorgenerator/requirements.txt
kaleido==0.0.2
python-dateutil==2.8.1
numpy
pandas==1.1.3
plotly==4.9.0
requests==2.27.1
pymongo==3.10.1
motor==2.1.0
Jinja2==2.11.1
beautifulsoup4==4.9.1
lxml==4.6.5
Pillow==9.0.1
markupsafe==2.0.1
beautifulsoup4==4.9.1